tests/resources/tree-construction/whitespace.dat

#data
  foo bar baz
#errors
#document
| <html>
|   <head>
|   <body>
|     "foo bar baz"

#data
<!--foo--> bar baz
#errors
#document
| <!-- foo -->
| <html>
|   <head>
|   <body>
|     "bar baz"

#data
  foo <!--bar--> baz
#errors
#document
| <html>
|   <head>
|   <body>
|     "foo "
|     <!-- bar -->
|     " baz"
